refactor: 패키지 이동 및,에러 메세지 하드코딩 제거
