+++ src/main/java/domain/Way.java
@@ -1,4 +1,4 @@
-package util;
+package domain;
 
 import java.util.Arrays;
